+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add `Table` and `TableWrapper` and `hooks/useScrollEdgeDetect`
 - feat: Add basic `hooks`
 - chore: Initial scaffolding